Slices

Slices from a MProb

MomentOpt.doSlices — Function.

doSlices(m::MProb,npoints::Int,parallel=false)

Computes Slices of an MProb

source

MomentOpt.Slice — Type.

Slice

A slice in dimension $j$ of a function $f \in \mathbb{R}^N$ is defined as $f(p[1],...,p[j],...,p[N])$, where p is the initial parameter vector and p[j] = range(lower[j], stop = upper[j], length = npoints), where lower[j],upper[j] are the bounds of the parameter space in dimension $j$.

Fields

res: Dict of resulting slices. For each parameter $p_j$ there is a Dict with as many entries as npoints chosen in doSlices

p0: Initial parameter vector dict

m0: data moments dict

Examples

julia> using MOpt
julia> m = MProb()
julia> p = OrderedDict(:p1=>1.1,:p2=>pi)
julia> m = OrderedDict(:m1=>rand(),:m2=>e)
julia> Slice(p,m)

source

MomentOpt.optSlices — Method.

optSlices(m::MProb,npoints::Int,parallel=false)

Computes Slices of an MProb and keeps the best value from each slice. This implements a naive form of cyclic coordinate descent in that it optimizes wrt to one direction at a time, keeping the best value. It's naive because it does a grid search in that direction (and disregards any gradient information). The grid size shrinks, however, at a rate update.

source